+++ changelog.md
@@ -7,6 +7,7 @@
 - Use Roboto font
 - Bump python version to 3.12
 - Fix showing the elements with no css styles
+- Rename the project to personal-portfolio
 
 ## v0.1.0